factor out lichess.trans
