+++ ui/site/src/index.js
@@ -1,3 +1,4 @@
 require('./util');
+require('./trans');
 require('./socket');
 require('./main');
